Flatten UhApicState
